Added contact deletion confimation dialog #150

+++ app/src/main/java/com/chooloo/www/koler/ui/contact/ContactFragment.kt
@@ -6,6 +6,8 @@ import android.view.LayoutInflater
 import android.view.View.GONE
 import android.view.View.VISIBLE
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
+import com.chooloo.www.koler.R
 import com.chooloo.www.koler.databinding.FragmentContactBinding
 import com.chooloo.www.koler.ui.base.BaseFragment
 import com.chooloo.www.koler.ui.phones.PhonesFragment
@@ -99,8 +101,16 @@ class ContactFragment : BaseFragment(), ContactContract.View {
     }
 
     override fun deleteContact() {
-        _activity.deleteContact(_contact.id)
-        finish()
+        AlertDialog.Builder(_activity)
+            .setCancelable(true)
+            .setTitle(getString(R.string.warning_delete_contact))
+            .setPositiveButton(getString(R.string.action_yes)) { _, _ ->
+                _activity.deleteContact(_contact.id)
+                finish()
+            }
+            .setNegativeButton(getString(R.string.action_no)) { _, _ -> }
+            .create()
+            .show()
     }
 
     override fun setFavorite(isFavorite: Boolean) {
